Moved output string creation to sep. files
Created get-output.js and get-output-standard.js (and tests).
get-output-standard.js does the actual work (and more files
like this will com). get-output.js deligates to this (these)
files according to what Error is to be printed.

Also created a skeleton file with logging function that
doesn't do anything yet. TODO.

5.30 AM and I'm so tired. Not sure if I'm finished with
parse-error.js file, which was the goal of this commit.
I commit anyway since I have no errors in my tests and
the program runs without error. And at a quick glance of
the file it seems to be doing what is should. I'll look closer
at the file when I wake up tomorrow.