Removed 'props' arg. in AbstractError
Until now, I had coded the Abstract class and the custom Error classes
thinking that I somehow needed to pass the name of the child Error class
to the parent. First I tried passing the child instance itself as a
2nd argument to the parent constructor, via super(). When I, after much
effort, realized that my tests were failing because that was not allowed in
JavaScript ('this' can't be used before super() and not either passed
as an argument to super()).
After that, I passed an object with a 'name' property with the child class'
name as a 2nd argument. And now, much later, I just realized that all this
was unnecessary because 'this' inside the parent object actually refers
to the child object from which the constructor was called with super().
Lesson learned... the hard way :/