Kasper for validating objects against JSON schemas
By using Kasper package for validating documents and other objects
I didn't need the 'validate-field.js' and 'BSON-JS-map.json' files
any longer and they were deleted. I validate the fields in documents
now in file 'validate-documents.js' with Kasper.

I also validate task objects in 'main.js' with help of Kasper and the
new schema files 'task-schemas.js' and 'table-scraper-task.js'.

Moved 'table-scraper.js' from lib/ to scrapers/
The 'scrapers/index.js' file is an object for linking scrape functions
with scraper names, used by 'create-tasks.js' when building
task objects.

In this commit I planned on doing a test file or more, but never
finished it because of the implementation of Kasper. The test file
I started but haven't finished 'test-create-tasks.js'.

File 'task-schemas.js' is a file for linking scraper names with
the corresponding schema for the task of that scraper. used
in 'main.js'.